\name{length}
\alias{length}
\title{length method for \code{Signal} class}
\description{Return how many listerns a \code{Signal} object contain}
\value{Numieric value indicate how many listeners this signal contains}
\docType{methods}
\seealso{\code{\link{listeners}}}
\author{Tengfei Yin}
\arguments{\item{x}{\code{Signal} object}}
